[examples][fidl] Update echo_client_go to connect to a server by URL
This patch makes Go echo client to take a server URL as an arguemnt
and connect to it instead of discovering the service in the environment.
This new behavior is in line with the echo clients for other languages.

Two command-line options are available (--server also works):

$ echo_client_go -h
Usage of echo_client_go:
  -m string
        message (default "Hello, Go World")
  -server string
        server URL (default "echo_server_go")

TO-839

Change-Id: Id80e41a840a0dce5a53d0a33f4b30a83ee495d46