[debugger] Implement continue.
Suspends all threads when an exception is hit, and allows resuming from
the exception with the "continue" command.

Fixes a bug with thread shutdown where all of the information sent in
the notification was invalid because the thread doesn't exist any more.

Moved OnThreadStarting/Exiting notifications from the Process to the
ProcessImpl in the client model (this doesn't need to be part of the
public API).

Change-Id: Ib6347e534248f070f5ddd2e9c35aef564e9b71f5