[zxcrypt] Use wait_for_driver_bind in TestDevice
This CL replaces the block watcher from TestDevice with the more
standard one in ulib/fs-management.  The most salient difference is the
lack of a timeout, but that's acceptable as these routines are not
expected to fail.  It also adds error messages around the calls to bind
drivers to devices, so that the culprit may be more easily identified
in the case of test flake.

ZX-1825 #comment Expected to be fixed

Change-Id: I0caaa9bb752492ebba84a3def90eb46df86bcabc