[modular][resolver] Avoid churning context suggestions when context changes at a high rate.
This is a slight hack, whereby we cache all the Daisies (the most
unique, but also clone-able struct) that describes each proposal and
compare them before clearing and re-issuing the same exact set of
proposals.

This is strong support for changing the entire Next space data flow
model to be pull, not push, as Next space knows exactly when it needs
suggestions and can cause agents to issue a single context update at
that time. We are now doing a ton of work in the background that never
is displayed to the user.

Change-Id: I54be26534f00e7596e4d518984da25668ea51cb2