[scenic] Add other options for SCENE_MANAGER_VULKAN_SWAPCHAIN
On the hikey there's no display, so a fixed size swapchain is the only
option. However on the VIM2 the swapchain size needs to match the size
of the display, so it should still be queried. We can add multiple
options to scene_manager_vulkan_swapchain to make this work.

Change-Id: I1b23123b728101aed1b99a82f361f21142187469