[fidl] Record sizes in coded types
Coding tables for vectors need to know the element size for the vector.
In most cases, this size is known from the flat::Type. However, when
that type is an identifier (which has size 0), we must perform a lookup
to find the coded type for the identifier. Storing the size in the coded
type gives us the proper element size for such a vector.

TO-898 #done

Change-Id: I66213c0c910f426dd1858f24172c8f780a6c6cb2