[gn] Make fuchsia_packages build argument a list of strings
This is the natural syntax to use in GN, and is also JSON
encoding for the same thing.  Scripts generating `--args=...`
will have to adapt to produce the JSON/GN list syntax instead of
a comma-separated list with no other punctuation.  The scripts
run by GN now take their `--packages` switch in the JSON format.

Change-Id: I1580bd1ca23297494db7607d92b2e74a026918df