[gn] Support shortcut selectors in `select_variant`
Add `select_variant_shortcuts`, defining a set of shortcut names.
These can be used in `select_variant` by making an element that's
a string instead of a full scope.  The string matches one of the
shortcuts exactly or is "shortcut/output_name".  The shortcut
expands to a list of full selectors (scopes).  Each variant name
is available as a shortcut, and then there is a special shortcut
for "host_asan" that expands into a standard blacklist+whitelist.

This makes selecting variants with a hand-written `select_variant`
GN build argument about as concise as the `gen.py --variant`
syntax, so `gen.py` no longer provides anything nontrivial.

TO-847 #comment Support shortcut selectors in `select_variant`

Change-Id: Ide32ad8974560f3926c8e903f74b797942a20df2