[suggestion_engine] Update feature context directly from SuggestionEngineImpl
This fixes an issue reported by larrylandry@ in which suggestions
were not appearing for queries and a fatal error was appearing in the logs.

I introduced the issue in https://fuchsia-review.googlesource.com/c/peridot/+/122439 when introducing a way
for ranking features to request a piece of context they needed to perform their ranking.
The reason of the breakage is in QueryProcessor's RankedSuggestionList instance.
QueryProcessor doesn't have access to the latest context so it was requesting a context
value through TakeContextValue and passing a nullptr to it. Further testing would have
avoided this issue as well as better API design (don't default arguments to nullptr, and
don't propagate information too many levels down).

In this CL I make sure that ranking features have the context they need and that context
is updated by SuggestionEngineImpl (who initializes the ranking features and also listens to
context updates) instead of RankedSuggestionList when ranking. This way every time we rerank
we know that the ranking features will have the context they need.

Change-Id: Iffcb7a412f07f14911d804c63874d7c9f0faebb4