[runtests] Add garnet test defaults, don't exit on all failures
This changes runtests to look at test directories for both zircon
and garnet, and not consider it a fatal error if a given directory
does not exist. It may be possible to better determine where to
look for tests via directory traversal or stat checks, but it would
increase the scope and complexity of this tool.

Change-Id: I370a0db8c85a9f31e63c985e7edcbf8cbb9b38c1