[dev][sdmmc] use BTI
This is a re-land of I1cb6db88c8a08c88c0de811acdb8123936fc9ffa (git
5a04ac31d589785303c0f41913fd59dbd5ad46ef), which was reverted in
335c3a414e169804eb9050d784af6cebd0f5c77f.  Testing on Acer and Paradise
showed no issue.  The cause of the revert was this change making
an existing buffer overrrun in the SDMMC driver become more visible (ZX-1851).
The only changes are additional logging in error cases.

Change-Id: Ia078bf659793983fc9240b0de1355e1d4d4d5c7c