[fidl][go] Generate interface proxies
This change primarily adds support for generating interface client-side
code, and tests for generating as well.

It also makes some changes to fix the IR tests, notably not to generate
an empty entry for a nullable handle in the struct tag and updating the
tests to work with the recent identifier changes.

TO-324

Change-Id: I52e0fc0ba87999d53cfa928942d16b2335585d78