[audio][mixer] Add multi-frequency set
This CL formalizes a set of frequencies to be used for audio
frequency response testing, and other upcoming fidelity tests.

Working in concert with the previously-added GenerateCosine
function, these summary frequencies (currently 40 Hz, 1 kHz
and 12 kHz) are "phase-locked" to the size of the buffer used
in our frequency-based testing. Also, we adjust the values
slightly so that periods do not closely relate geometrically
to sample rate -- so that sampling of a given sinusoid is
more statistically spread across the entire waveform (e.g.
approximately 11997.8 Hz instead of 12000 Hz).

For now we assume an eventual 48 kHz output sample rate, so
(along with source buffer of size 65536) our translation from
'freq' array value, to effective frequency within our buffer,
is a ratio of 65536/48000. In other words, to simulate a 1 kHz
waveform, we would send GenerateCosine a 'freq' value of 1363.

Change-Id: I7776e9e23daae507c2fdee7b7970634c67b8a94c