[fs] Ensure Vnode::Close is called on unclean teardown
The contract of ulib/fs is that every call to "Open"
will be paired with an accompanying call to "Close",
in the event of either:

1) A client sends a close message explicitly, making
their channel defunct (and no longer observed), or
2) A client closes their channel, implicitly closing
their connection to the filesystem.

Although the Connection class in ulib/fs would properly
cancel waits and tear down the connection object,
it previoulsy missed sending a close in the second
condition, since it invoked "CallHandler" while the
channel was not yet ZX_HANDLE_INVALID.

This patch ensures that the close handler is invoked
in this condition.

ZX-1853 #done

Change-Id: I291192cae32e8250f7cef932cbf1a4de4b8b2d2d