[media] replace NetMediaPlayer with MediaPlayer
This CL updates net_media_service so that remoting can happen using
the MediaPlayer interface instead of NetMediaPlayer. The
NetMediaPlayer code remains until higher layers no longer reference
it.

NetMediaPlayer was defined because MediaPlayer didn't have a way to
tell it to open a URL (a SeekingReader had to be provided, something
that can't happen over the wire). Now MediaPlayer has a SetHttpSource
method, which is sufficient (it doesn't support file URLs, but that's
OK for now).

Prior to this CL, a client that wants to control a remote player
needed to use the NetMediaPlayer interface. A proxy exposing this
interface could be created using
NetMediaService.CreateNetMediaPlayerProxy. A party wanting to
expose a player for remote control had to wrap a MediaPlayer as a
NetMediaPlayer by calling NetMediaService.CreateNetMediaPlayer.

Now clients wanting to control a remote player can use MediaPlayer,
which they get from NetMediaService.CreateMediaPlayerProxy. A party
wanting to expose a player for remote control just publishes it
using NetMediaService.PublishMediaPlayer.

Once Topaz is using the new NetMediaService methods, the old ones
(and NetMediaPlayer) will be removed.

Change-Id: I2353e2840ad0fcd04977efa638439e202389a869