[media] rename net message types in NetMediaService
The NetMediaPlayer interface will soon be removed. The MediaPlayer
interface can now handle remote scenarios (because it has the method
SetHttpSource), so NetMediaPlayer is no longer needed.
NetMediaService will be modified so it no longer refers to
NetMediaPlayer.

This CL takes a first step by renaming the messages used on the wire
for media player remoting. The term 'NetMediaPlayer' is replaced by
the term 'MediaPlayer' and files moved accordingly.

Change-Id: Ib7004c30fd025286d5aa063d5b8f2833e22ed216