[vmo_mapper] Added an offset option to VmoMapper
This change adds the argument 'offset' to VmoMapper.
The offset is used directly in the vmar mapping function.
All the modules that depend on VmoMapper are being updated;
the intel-hda-controller in zircon, and the audio_renderer2
in Garnet.
Since this change breaks Garnet, there will be another CL
to unstick the roller in Garnet.

Change-Id: Iadfe39a57ccc9209098376c3cb31966ba92b89ec