[media] allow media players to have multiple clients
This CL adds the method MediaPlayer.AddBinding, which allows a client
to duplicate a binding to the player. This enables multipoint
control of the player. This has many potential uses, but is currently
motivated by the the 'fling' scenario in which both a local client
and a remote client want to control a player. Previously, this was
done by wrapping the player in a NetMediaPlayer, which lived in
NetMediaService, where the stubs for remoting also lived. In the
future, we will no longer wrap MediaPlayers but rather add a new
binding to them and then publish one of the bindings via
NetMediaService.

Change-Id: I3dce7572c1807a2e6201b6019d694491fe719248