[audio][mixer] Move noise floor tests to new file
This CL moves the noise floor tests out of the transparency
grouping (which previously meant no-gain-no-SRC, but now
means bit-for-bit testing), into a new grouping of tests
labelled frequency_tests. The noise floor tests are only
the first of these, as they will include the incoming
frequency response, signal-to-noise and dynamic-range tests
as well as potentially phase and impulse tests in the
future.

Change-Id: I1251a7d9d50c41b442aa85e74a4c7bd25db405e4