[push-package] Trigger on-device module_resolver to index a pushed package.
`fx push-package` should now try to trigger the module_runner to index
the module manifest inside the package if it exists. Note that this only
works if a user is logged in (since module_runner runs under a user).

Putting this trigger logic inside push-package is not the final goal and
will go away altogether once we have more infrastructure in place, but
for now will eleviate a lot of development burden that the upper layers
face when doing module development.

MI4-897

Change-Id: I16f63cc2c115572ac1f776e2dfdb8976f31e3f5c