[zircon] Package up implementation and ABI SOs for prebuilt libraries
This packages up the ABI shared object file and the implementation
shared object file for shared libraries exported from Zircon as
prebuilts.  The ABI SO contains information for linking, used at
compilation time, but doesn't include the implementation of functions.
The implementation SO contains the actual functionality and is needed
at packaging time.  The layout for SDK artifacts is:

arch/{x64|arm64}/lib/libfoo.so <- ABI SO, for linking
arch/{x64|arm64}/package/libfoo.so <- Implementation SO, for packaging

This uses additional information exported from Zircon in
https://fuchsia-review.googlesource.com/c/zircon/+/135002, but works
for older versions of Zircon as well so that the change can be rolled
up through each layer.

Change-Id: I54c05a3b6f34dccec9499e4c8211c36d072f7c7a
TO-918: #comment