[fvm] Allow transactions over maximum size.
Since the FVM layer previously did not split transactions based
on the maximum size specified by the underlying driver, it was
possible in some cases (specifically, when trying to load an FVM
partition with metadata over a particular size), for the metadata
read from disk to fail.

This patch splits up outgoing transactions from the FVM layer into
a number of smaller transactions, similar to what the BlockServer
is already doing.

ZX-1877 #done

Change-Id: I1799d16999d3584a64033e24ffcc22b298e31545