[debugger] Clarify process exit and detach.
Renames the "stopped" state to "none" (in code) and "not running" (in
the status messages). Previously it woudl say "Stopped" which was
misleading.

Updates the TargetObserver to include a process create and destroy
function rather than one state change function. This is much more clear
and allows adding the exit code and process destuction reason on the
destroy notification.

Updates messaging around process detaching.

Change-Id: I1e46e79617eecc910f9b5ef3b01507f758ddaf5c