[dart_test] Fix build label for Flutter test shell
The label at //third_party/flutter/shell is a helper that depends on
the actual platform-specific tester shell as well as other targets that
aren't needed by the shell, like //third_party/flutter/shell/embedder.
This makes the test target label more specific, so we only build the
code we need.

Change-Id: Ice839cf3b5a7e0b0db19c40b1eecf59d302e3730