[audio][mixer] User can specify the resampler type
An additional, optional parameter to Mixer::Select can be used
to specify the type of Resampler to be used in the requested
mixer. An enum of possible resampler types, including Default,
is found in mixer.h.  If this parameter is absent, or if
Default is specified, then the existing (default) resampler
selection algorithm is used.  Otherwise, if a specific
resampler is specified, but that resampler is not workable for
the configuration provided, then nullptr will be returned,
even if a different Resampler could have been successfully
instantiated.  Care should be taken, with Default specified
whenever possible.

MTWN-90 #done

Change-Id: I73900d941b30bf96d93de90d09ea7521a26beba7