[debugger] Start on breakpoint backend.
First pass at implementing backend breakpoints. This implementation is
still very raw, and breakpoints can not be stepped over yet. Currently
working is the ability to set, enable, disable, and remove
process-specific breakpoints on specific addresses.

When these breakpoints are hit, the process will report that it's
suspended.

The debug agent needs to track more state about why a thread is stopped
and how to resume it, so a new class in the debug agent is created to
represent a thread. Some code moved here from the DebugAgent class. It
seems like the debug_agent ExceptionHandler needs a significant refactor
to make this more clean (there is a lot of plumbing that should be
reduced), which I'll do in a followup.

Change-Id: Ia6dec7a174131922f55c362eade5b6bbd57e82ae