[wlan] Explicitly initialize class' fields
For some reasons, device.h needs an explicit initialization of its
parent, zxdev and wlanmac_dev to work with my Pixelbook. Without this
explicit initialization the interface cannot be created due to
'wlanmac_dev' being not a nullptr.

Same issue happens with the BeaconSender when enabling finspect.

NET-627 #done #comment

Change-Id: I50a6648657138dc03be7309da6f880473e017d9f