[machina][virtio] Support VIRTIO_F_EVENT_IDX.
Add support for VIRTIO_F_EVENT_IDX feature to support more efficient
interrupt & notification suppression for some devices.

If this feature is enabled, the driver is able to write a queue index
to the 'used event' field in the ring that tells the device to not send
any interrupts until that many buffers are available in the used ring.

Similarly, the device is able to upate the 'avail event' field in the
avail ring to instruct the driver that ring kicks can be elided until
the avail index reaches the desired value.

This feture can allow for more efficient processing of virtio-net queue
buffers. The driver can indicate that reclaming of used TX buffers can
be batched up instead of taking one interrupt for every buffer.
Likewise the device can do similar to reduce ring kicks for buffers
added to the RX queue.

This also removes the need to call VirtioDevice::NotifyGuest after
calling VirtioQueue::Return, as we're enabling VirtioQueue::Return to
handle interrupt suppression via virtio standard methods.

PD-96 #done

Change-Id: Icc446c937da3a554dd8a9f7e02261cda8d97acf0