[wlan][ralink] Revamp USB outbound data structure
Rename struct TxPacket to struct BulkoutAggregation.
This reflects the intendion of this struct, which is
to assist bulkout transfer of the outbound frames.

BulkoutAggregation's structure is better studied, and its header,
payload, and tail padding are explicitly commented for
readers' future references.

The structure and the use of TxInfo leaves some room to study
further. That said, Fuchsia's use is to serve as the header
4 bytes of BulkoutAggregation. There its payload length field
is renamed for clarification of that field's use.

NET-459 #comment

Change-Id: Iada792c9febe13cd100a3d66739e753fa17bb716