[arm64][build][kernel] arm64 kernel is now a bootdata container
Like the x64 kernel, the arm64 kernel is now packaged as a bootdata container.

Currently, all of our arm64 platforms except the hypervisor do not support this.
For those platforms we now have a shim that is prepended to the beginning of the kernel.
This shim checks to see if there are remaining bootdata items following the kernel.
If there are, the shim copies the kernel beyond the end of the bootdata
and executes executes it there. Otherwise it will jump to the kernel in place.

We also add a new <board>-combo-bootdata.bin build file, which contains
the kernel and all of the bootdata packaged in a single file.
This is now being used on the vim, vim2 and hikey960 targets.
However we cannot use the combo-bootdata on gauss, qemu and odroidc2,
so on those platforms we add the shim to their kernel and boot with a
separate bootdata passed as the ramdisk.

In this change we also add --header and --header-align options to mkbootfs,
which prepend an optional binary header to the beginning of the mkbootfs output.
Strictly speaking we could have done this using "cat" instead,
but having this option in mkbootfs eliminates the need for temporary files
and would be simpler to use in the Fuchsia build system.

Change-Id: I9fa572318158043e7b22c5ad944f33a23053abd3