[uart] Use AUTOUNSIGNAL for Tx (FIFO full) event.
The uart Tx event is intended to allow blocking Tx to wait until the
Tx FIFO empties. Blocking uart Tx will spin if we don't either explicitly
unsignal the event or use AUTOUNSIGNAL (easier). This change makes the
event AUTOUNSIGNAL (and ensures that we unmask the Tx interrupt before
event_wait()'ing).

Bug: ZX-1883

Test: Added a switchcount to the thread struct confirming that we block
in event_wait() (when the Tx FIFO is full).

Change-Id: Ieaf796e0dc665380c1b91beb982033799c61690c