[fidl] Improve table generation for unions
We always need to generate a coding table for unions because we need to
validate the tag value. Also, we need to generate an entry for each
member of the union because we use the index of the member in the array
to associate the member with the tag value.

Change-Id: I49428c37e1e0bca6abae3025d902f656ffc5b62a