[media] Stop media_service_tests from crashing
Previously, these tests relied on a compiler optimization to coalesse
short string literals. Compilers are not required to perform this
optimization. Now, we use strcmp to check that the strings are equal
rather than simply checking the the address of their buffers.

Also, convert the declarations to use char[] rather than char* for fewer
relocations.

Change-Id: Icd53c66b290f98401c73dab8cb0202f377b4bc03