[ulib][ethernet] Fix a potential deadlock in eth_get_buffer_locked
The eth_get_buffer_locked function attempts to get a TX buffer from the
pool. If none is available, it tries to reap buffers by calling
eth_complete_tx, which in turn does a non-blocking read from tx_fifo.
If the read from the fifo returns SHOULD_WAIT and 'block' is set to true,
the code currently performs a wait on the rx_fifo, which I believe is a
typo. RX buffers are never returned to the pool; instead, they are
always immediately enqueued for RX, so it doesn't make sense to wait on
the RX fifo in order to get a TX buffer.

Change-Id: I691c10656e56eb756833cee84b0ca1e93274b6ca