[debugger] Continue from breakpoints on x86
Allows continuation from breakpoints by removing the breakpoint
instruction, single-stepping the thread, and then putting the breakpoint
instruction back. This introduces a race condition with other threads
during the "breakpoint removed" window, but should be good enough for
now.

This works on x86 but not ARM yet. There is a kernel bug around
single-stepping that causes problems in some cases, and the addresses aren't
quite right for single-stepping (the kernel bug makes this hard to diagnose).

Change-Id: I534b0952c58361123602f576fb784199a0f53c15