[auth] Fix dev_token_mgr_e2e_test
Previously, these lambdas captured the tokens variable by reference,
which is problematic because that variable is stored on the stack.
Previously, this test must have pasted by good luck in that the freed
stack must not have been touched between the time we sent the request
and the time we got the response.

Now, we capture the tokens variable by value.

Change-Id: I2e072ba9faf885b36dcfc7afb233c9842e4231d1