[fidl][go] Ensure channel is closed on error
This change makes sure that the underlying channel is closed if any
server error is encountered, so that the client (whose calls are
asynchronous) can make progress in the event of an error. Before, this
was prevented because CancelWait would fail in an expected manner (i.e.
the wait ID was not found since the wait ID is removed from the
dispatcher's bookkeeping when calling into the handler).

FIDL-103

Change-Id: I1cf95e6a8bb6dc3295219b55c47d4f5cd31aa4b8