[ledger][p2p] Coordination layer for Cloud and P2P sync
Creates a coordination layer between Storage, cloud_sync and p2p_sync.
This CL is mostly a refactoring as no behavior change is introduced:
only the Cloud is used to upload and download objects.

This change will allow us to enable p2p upload and download of objects
and commits.

LE-446 #done

Change-Id: I2500655ad73f43857b6b16ce8253b7c9d3fd009a