[ledger] Remove launch_benchmark tool.
Since `trace record` with a spec file is now a standard way to run
trace-based benchmarks (as described in
https://fuchsia.googlesource.com/garnet/+/master/docs/benchmarking.md),
launch_benchmark tool is not that useful anymore, being just another
way to run the same benchmarks.

LE-440 #comment Some of the recent changes broke the tool; but given
that it's not widely used right now and there are other ways to run
Ledger benchmarks, I lean towards removing it entirely.

Change-Id: Id7284aa64385e4bf112b58f86c64eaa4b4197629