[module_manifest][gn] copy_module_manifest() rule now depends on module_manifest_package()
This moves module_manifest_package() into module_manifest.gni. The
intention that no one is calling module_manifest_package() manually in
the upper layers, but it is done automatically as part of flutter_app().

This should make it so flutter_app() rule (as part of topaz) automatically
generates module_manifest_package() along with storing a copy of the
manifest in $build_dir/module_manifests/

Following this patch are patches in the upper layers to stop calling
module_manifest_package().

Change-Id: I4d4069164e0e1855d1fb9b580bb3af1a6d23fd40