[audio][ihda] Fix a bug in the PinnedVmo page merger.
The code which was attempting to merge adjacent chunks of memory in
the PinnedVmo class had a bookkeeping error during it's preflight
phase while attempting to compute how much storage was needed to hold
the final region list.

By failing to update "last", the code would only ever think that there
was at most 1 region which could be merged.  It would then allocate
storage for either N or N-1 (where N is the number of pages in the
VMO), but only fill out entries which could be successfully merged,
and leave the remaining entries un-initialized.

Properly updating "last" during the preflight loop avoids this mistake
and causes us to allocate a tightly packed region array, as we should.

Change-Id: I6d055acef46e7e325637197e3ad8cdf6d9ee7a50