[wlan][mlme] Fix message serialization
Subtracting off the size of the ServiceHeader was an error if the
encoded size of the message was smaller than ServiceHeader bytes in the
first place. Rather than trying to trick the encoder, we stop using our
own ServiceHeader and let it encode the fidl_message_header_t for us.

NET-647 #done

Change-Id: I28ecef8ed14f4238485bc9c58e9591f24f0bf7ef