[audio][mixer] Tighten up mixer optimization
An optimization to the audio mixer skips mixing operations if the
gain is below a threshold at which all output will be zero. This
CL fixes an off-by-one comparison, pulls the optimization up one
level in the capture case, and adds static_asserts so that the
Mixer objects enforce that they must be called with parameters
consistent with this optimization.

MTWN-76 #done

Change-Id: I12be056c92b3a86166e83a2c8bd18ebebb7ea250