[netstack] Restore hotfix for NET-298.
It turns out that there's a DHCP bug in third_party/netstack that
prevents multiple DHCP clients from concurrently listening on the same
port.

To reproduce, allow the first NIC (presumably wired) to fail to obtain a
DHCP lease, and then attach a second NIC (presumably wireless) that is
expected to successfully obtain a DHCP lease.

Details of the investigation in NET-298.

Change-Id: I0391c4db75f59c0cfee40391b061456253961010