[fs] Add "FLAG_EXACT" to mmap, re-enable MAP_SHARED for blobfs
- MAP_PRIVATE implies a COW clone of the underlying VMO must
be used.
- MAP_SHARED specifies that updates to the underlying file
should be visibile to the filesystem and vice versa.

These two modes are not capable of expressing a third modality:
acquisition of an 'exact', non-cloned VMO. Although MAP_SHARED
is superficially somewhat similar to this requirement, MAP_SHARED,
according to the posix specifications, only dictates the behavior
of updates, not the implementation. As a consequence, for read-only
filesystems, "MAP_SHARED" and "MAP_PRIVATE" can both be expressed
by requesting a clone of the underlying file VMO.

This patch expands the FDIO_MMAP flags to include "FLAG_EXACT",
allowing callers to express a desire for this 'exact match' which
is not expressed with MAP_SHARED.

PKG-31 #done

Change-Id: I564f91761ea7a8a17634de66eb3b004a2b0c9047