[audio][mixer] Round gain scalar during dB calculation
During the volume scaling process in the core audio mixer, we
take in two float values, for Renderer and Master gains, combine
them to form a single dB gain, then convert that into a fixed-
point multiplier (where an incoming 0.0 dB, representing no change
in volume, is represented by a '1.0' multiplier). In calculating
this fixed-point multiplier (in 4.28 form), we currently truncate
any fractional value beyond 28 bits of precisionn. With this CL, we
round, leading to greater gain precision.

MTWN-73 #done

Change-Id: I13d03e8df303653b547547f1c594e7648a854085