[audio][mixer] Round audio data during scaling
When scaling audio data, we should round instead of truncate.
With this CL, we round the data. With Subsequent CL (to fix
MTWN-73), we also round when formulating the gain_scale value
itself, eliminating bias and increasing accuracy further.

This CL addresses our sensitivity to gain. Primarily this is
shown in an improvement of about 4 dB in our "noise in the
presence of signal" numbers, clasically labelled as Dynamic
Range.

MTWN-80 #done

Change-Id: I22e63cd9e5c44477c6ebfae25c8560740362dda9