[media] add HttpReader and move FidlReader
This CL adds a HttpReader that implements the Reader abstract class.
This code is adapted from media_service/network_reader_impl.*, which
implements an http reader FIDL component that exposes SeekingReader.
This new reader will be used by the in-proc player.

This CL also moves FidlReader from bin/media/fidl to bin/media/demux.
This avoids a dependency cycle and puts FidlReader with the other
in-proc readers. FidlReader reads from a SeekingReader and implements
Reader.

Change-Id: I1a491a3431aba093c24c56c6f59c35fd4566847d