[media] move ConversionPipelineBuilder to player directory
ConversionPipelineBuilder is old code that used to be used for the
old in-proc player and has been kept around since then, built but
not used.

The new in-proc player will require ConversionPipelineBuilder. Most
of the code will go in bin/media/player, which can't depend on
bin/media/media_service, so ConversionPipelineBuilder needs to move
to bin/media/player.

Change-Id: I7ec66403d02ba13f05365481e7db7654003b4fab