[audio][mixer] Enforce that source buffs have integral length
Although the number of incoming source frames is expressed fractionally
(in a fixed-point 20.12 format), the actual number of frames is always
an integer. Because closely-related parameters (step size and source
offset) are also provided in this format, it is more convenient for
callers to keep the parameter in this format.

This CL enforces this requirement, with DCHECK. It also reworks the
resampler tests accordingly.

MTWN-77 #done

Change-Id: I2d32ce9571851a368f4202502cfc03e2f0d03c22