[arm64][machina] Add machina board definition.
The hypervisor had been using the qemu board definition until now. In
order to support the PCI interrupt mapping used by the hypervisor we
need to customize the PCI interrupt LUT.

This is largely a fork of the qemu platform with some unused
definitions removed and the PCI LUT swizzle logic updated. The PCI PIO
region has also been removed as this is not emulated by the hypervisor.

Change-Id: I09eb6e0a4e14fe6e06b37d6cea36d7d3fcf1f478