[fidl][go] Add support for unions in the generator
This change adds support for generating code for FIDL unions in the
golang generator backend. Unions are represented as golang structs
where every union member is just another field. There are two addtional
fields in these structs: the tag field, and an unexported dummy field
which marks the struct as a FIDL union to the encoder/decoder.

FIDL-97

Change-Id: I90cad498fb2e6aa1a01511536fb3c09cde4a6242