[debugger] Small improvements in console interface.
Fixes a bug where listing breakpoints and threads was impossible if
there was no current breakpoint or thread. The error checking was a
little too strict. If the current active thread or breakpoint was
deleted, it will prevent listing threads and breakpoints.

Fixes a bug in tab completion where if there was more than one token,
the space after the last one would be omitted. A test is added for this
case.

Requires that the "new" verb take an explicit "process" noun (i.e.
"process new" rather than just "new"). This is because I want to be able
to use "new" to maky any creatable noun in a standard way, and I don't
want people to get used to just typing "new" for the process case before
any other variants are implemented.

Change-Id: I9efc4f551466cfbe51ccf48d1006e7fc338d9754