[async][ledger] Refactor Environment for inclusion of async_t*.
-Have Environment initialize with an async_t*.
-Remove GetIORunner() considerations, as it is dead code.

This is an intermediate change in the effort to deprecate
fxl::TaskRunner.
After this change, once all classes that are constructed with
Environment are
refactored to use async_t* instead, the TaskRunner here will be removed.

Note: the default value of nullptr is a stop-gap; it can be removed once
LedgerAppInstanceImpl is refactored.

BUG: ZX-1819

Change-Id: Iedb07817c516a9e162cee25af19f09ae4afa0b01