[fidl][rust] Fix off-by-one error in Rust clients
The slab of message interests is zero-based, while FIDL txids are
one-based. A previous patch attempted to shift the message interest id
to generate a txid, but missed a shift when the transaction was over.

To prevent future mistakes, new types are introduced to let the compiler
ensure we don't use the wrong id.

Change-Id: Iaf2fe70b5b706347a4e0b4da67c3f22424f5f836