[audio][mixer] Enable float output
Enable the mixer core to provide its final audio mix to "downstream"
parties in 32-bit IEEE-754 float format. For now, we simply convert
at the end of the pipeline from an internal fixed-point representation
into floating-point, but eventually we anticipate that the internal
format will be IEEE-754 32-bit float as well.

In keeping with the industry standard, values corresponding to 0dBFS
are represented by float values +1.0f and -1.0f.

Note: there is still work to be done (driver_output and other places)
before audio output devices can actually receive audio in this format.
This CL just enables it in the mixer core.

MTWN-54 #done

Change-Id: Ifb9d17b65e2eb111f39ce2971be5e6f312bd4a54