[view_manager] Pointer event coordinates shouldn't be NaN
Previously, we were creating an event path that was twice as long as
expected because we were initializing the path with N empty hits and
then appending the actual N hits. Now, we start at zero and just append
the N hits.

Also, clean up a case where we were moving from a variable twice.

Change-Id: Id1bf0207e2d8eb9ccf87a056e956e5ba0056ff30