[debugger] Separate out debug_agent IPC decode.
The IPC decode logic used to be in the DebugAgent class. This change
changes it to be a separate class with a virtual interface implemented
by DebugAgent.

The change helps to focus the functionality of the DebugAgent on the
higher-level handling of IPC requests and process exceptions. More
importantly, it will allow the back end of the IPC requests to be mocked
for testing the frontend in isolation.

Change-Id: Id92bb5870615963a1c003776ff384eed2c6ac863