[wlan][brcmfmac] Rename skbuff functions to netbuf
Rename / shift skbuff-related functions to Fuchsia style.
Rename "skcb" and "skbcb" to reflect the "workspace" name (formerly
"cb" field of skbuff).
NOT a full rename of "skb" - that would be almost 1000 lines.
NOT an implementation of the functions.
Does include a few code changes where functions were combined or
unnecessary.

Test: Builds.

NET-981

Change-Id: I44376e420ab77e439c8eb6359b626b1a316153c8